Moved shared logic tests into correct package
