+++ packages/shared/package.json
@@ -10,6 +10,7 @@
   },
   "scripts": {
     "build": "tsc",
-    "watch": "tsc -w"
+    "watch": "tsc -w",
+    "test": "jest"
   }
 }
